+++ package.json
@@ -17,7 +17,7 @@
     "coverage": "codecov",
     "readme": "rio init && rio init --private",
     "pub": "npm version patch --force && npm publish",
-    "oas": "rio oas --private"
+    "oas": "rio oas && rio oas --private"
   },
   "dependencies": {
     "commander": "^8.2.0"
