+++ src/utils/rc.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+
 function getRioRC(path) {
   if (path) {
     // eslint-disable-next-line
@@ -14,7 +16,21 @@ function formatEndpoint(route) {
   return parts;
 }
 
+function writeToFile(fileName, content) {
+  /* istanbul ignore next */
+  if (process.env.JEST_WORKER_ID === undefined) {
+    fs.writeFile(fileName, content, (err) => {
+      if (err) {
+        console.log(`Failed to write ${fileName} due to error ${err}`);
+        return;
+      }
+      console.log(`${fileName} was written successfully`);
+    });
+  }
+}
+
 module.exports = {
   getRioRC,
   formatEndpoint,
+  writeToFile,
 };
